Fixed improved cover behaviour
Previously, it incorrectly attempted to fetch the improved cover every time the album name, or the artist name changed. Now, it will only fetch the improved cover at the start, and only apply it if requested by the user. Furthermore, this allows for custom album name and artist name, without it affecting the improved cover fetching function